icons/d/requirements.txt
aiohttp
beautifulsoup4
cohere
edge_tts
fastapi[all]
groq
mtranslate
openai
pydub
pymongo
python-dotenv
pywhatkit
Requests
rich
starlette
uvicorn
websockets
googlesearch-python